+++ SDK/mediation/mobileads-unityads/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 All notable changes to UnityAds Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 1.5.0
+
+#### Added
+* Added support for UnityAds sdk version 3.6.0
+* Updated minimum supported UnityAds sdk version to 3.6.0
+* Added support for Yandex Mobile Ads SDK version 4.0.0
+* Updated minimum supported Yandex Mobile Ads SDK version to 4.0.0
+
 ## Version 1.4.0
 
 #### Added